Save profile statistics to the database
* Ensure tables are created on application start
* Change pp and accuracy to be stored as doubles internally (just to be consistent)
+ Save profile statistics (load on player creation, then update with new scores)

+++ OsuServer/API/Packets/Server/UserStatsPacket.cs
@@ -20,7 +20,7 @@ namespace OsuServer.API.Packets.Server
             binaryWriter.Write((byte) Player.Status.GameMode); // mode id
             binaryWriter.Write(Player.Status.MapID); // mnap id
             binaryWriter.Write(Player.Stats.Values.RankedScore); // ranked score
-            binaryWriter.Write(Player.Stats.Values.Accuracy); // acc
+            binaryWriter.Write((float) Player.Stats.Values.Accuracy); // acc
             binaryWriter.Write(Player.Stats.Values.Playcount); // playcount
             binaryWriter.Write(Player.Stats.Values.TotalScore); // total scotre
             binaryWriter.Write(Player.Stats.Values.Rank); // rank
